Fix printf format specifiers for better portability

+++ test/test_basic.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <inttypes.h>
 #include "fixed-point-math.h"
 
 /*
@@ -13,17 +14,17 @@ int main(void) {
     /* Test Q1.7 functions */
     printf("Testing Q1.7 format functions:\n");
     q1_7_t q7_result = q1_7_sin(0);
-    printf("  q1_7_sin(0) = %d\n", q7_result);
+    printf("  q1_7_sin(0) = %d\n", (int)q7_result);
     
     /* Test Q1.14 functions */
     printf("\nTesting Q1.14 format functions:\n");
     q1_14_t q14_result = q1_14_sin(0);
-    printf("  q1_14_sin(0) = %d\n", q14_result);
+    printf("  q1_14_sin(0) = %d\n", (int)q14_result);
     
     /* Test Q1.30 functions */
     printf("\nTesting Q1.30 format functions:\n");
     q1_30_t q30_result = q1_30_sin(0);
-    printf("  q1_30_sin(0) = %d\n", q30_result);
+    printf("  q1_30_sin(0) = %" PRId32 "\n", q30_result);
     
     printf("\nAll function templates are accessible.\n");
     printf("TODO: Implement actual function logic and comprehensive tests.\n");
